Major update - project finished
Followed an electron tutorial to create an app that requires a title and content, and once you hit the "Submit" button, the app will create a file with the given title and content. For now, the file location is in the current directory, but it will be changed later so that the user can choose the location.